# add `augment()` method for `model_stack`s