--- a/R/predict.R
+++ b/R/predict.R
@@ -224,13 +224,11 @@ augment.model_stack <- function(x, new_data, ...) {
   
   res <- dplyr::bind_cols(new_data, predict(x, new_data = new_data, ...))
   
-  if (isTRUE(dots[["members"]]) & 
-      x[["mode"]] %in% c("regression", "censored regression")) {
+  if (mode_is_regression(x) & isTRUE(dots[["members"]])) {
     res <- dplyr::rename_with(res, ~paste0(".pred_", .x), any_of(member_cols))
   }
   
-  if (x[["mode"]] %in% c("regression", "censored regression") &
-      outcome %in% colnames(new_data)) {
+  if (mode_is_regression(x) & outcome %in% colnames(new_data)) {
     res <- 
       dplyr::mutate(
         res,