Closes #173.

``` r
library(tidymodels)
library(stacks)
```

An example with the built-in `reg_*` tuning results:

``` r
reg_st <- 
  stacks() %>%
  add_candidates(reg_res_lr) %>%
  add_candidates(reg_res_svm) %>%
  add_candidates(reg_res_sp) %>%
  blend_predictions() %>%
  fit_members()

reg_st
#> ── A stacked ensemble model ─────────────────────────────────────
#> 
#> 
#> Out of 16 possible candidate members, the ensemble retained 3.
#> 
#> Penalty: 1e-06.
#> 
#> Mixture: 1.
#> 
#> 
#> The 3 highest weighted members are:
#> # A tibble: 3 × 3
#>   member          type         weight
#>   <chr>           <chr>         <dbl>
#> 1 reg_res_svm_1_3 svm_rbf    0.722   
#> 2 reg_res_sp_03_1 linear_reg 0.478   
#> 3 reg_res_sp_10_1 linear_reg 0.000465
```

A straightforward application:

``` r
augment(reg_st, tree_frogs)
#> # A tibble: 1,212 × 9
#>    clutch treatment  reflex    age t_o_d     hatched latency .pred .resid
#>    <fct>  <chr>      <fct>   <dbl> <chr>     <chr>     <dbl> <dbl>  <dbl>
#>  1 168    control    full   466965 morning   yes          22  33.5 -11.5 
#>  2 145    gentamicin full   404310 afternoon no           NA 168.   NA   
#>  3 149    gentamicin full   426220 night     no           NA 146.   NA   
#>  4 100    control    mid    355360 night     no           NA 128.   NA   
#>  5 230    gentamicin mid    356535 night     no           NA 167.   NA   
#>  6 99     control    low    361180 night     yes         360 186.  174.  
#>  7 145    gentamicin full   400070 afternoon no           NA 152.   NA   
#>  8 133    control    full   401595 afternoon yes         106 112.   -5.65
#>  9 100    control    mid    357810 night     yes         180 124.   55.5 
#> 10 182    control    mid    358410 night     no           NA 124.   NA   
#> # … with 1,202 more rows
```

Without the outcome, no `.resid` is present:

``` r
augment(reg_st, tree_frogs %>% select(-latency))
#> # A tibble: 1,212 × 7
#>    clutch treatment  reflex    age t_o_d     hatched .pred
#>    <fct>  <chr>      <fct>   <dbl> <chr>     <chr>   <dbl>
#>  1 168    control    full   466965 morning   yes      33.5
#>  2 145    gentamicin full   404310 afternoon no      168. 
#>  3 149    gentamicin full   426220 night     no      146. 
#>  4 100    control    mid    355360 night     no      128. 
#>  5 230    gentamicin mid    356535 night     no      167. 
#>  6 99     control    low    361180 night     yes     186. 
#>  7 145    gentamicin full   400070 afternoon no      152. 
#>  8 133    control    full   401595 afternoon yes     112. 
#>  9 100    control    mid    357810 night     yes     124. 
#> 10 182    control    mid    358410 night     no      124. 
#> # … with 1,202 more rows
```

Dots are passed to `predict.model_stack`:

``` r
augment(reg_st, tree_frogs, members = TRUE) %>%
  select(latency, starts_with("."))
#> # A tibble: 1,212 × 9
#>    latency .pred .pred_reg_res_…¹ .pred…² .pred…³ .resid .resi…⁴ .resi…⁵ .resi…⁶
#>      <dbl> <dbl>            <dbl>   <dbl>   <dbl>  <dbl>   <dbl>   <dbl>   <dbl>
#>  1      22  33.5             41.8    32.2    29.2 -11.5    -19.8  -10.2    -7.18
#>  2      NA 168.             135.    158.    171.   NA       NA     NA      NA   
#>  3      NA 146.             125.    136.    139.   NA       NA     NA      NA   
#>  4      NA 128.             112.    120.    122.   NA       NA     NA      NA   
#>  5      NA 167.             140.    166.    160.   NA       NA     NA      NA   
#>  6     360 186.             130.    235.    214.  174.     230.   125.    146.  
#>  7      NA 152.             137.    154.    134.   NA       NA     NA      NA   
#>  8     106 112.              95.5   112.    112.   -5.65    10.5   -6.08   -5.58
#>  9     180 124.             111.    125.    114.   55.5     68.5   55.4    65.7 
#> 10      NA 124.             111.    126.    114.   NA       NA     NA      NA   
#> # … with 1,202 more rows, and abbreviated variable names
#> #   ¹​.pred_reg_res_svm_1_3, ²​.pred_reg_res_sp_10_1, ³​.pred_reg_res_sp_03_1,
#> #   ⁴​.resid_reg_res_svm_1_3, ⁵​.resid_reg_res_sp_10_1, ⁶​.resid_reg_res_sp_03_1
```

Same goes for classification:

``` r
tree_frogs_class <- 
  tree_frogs %>%
  dplyr::select(-c(clutch, latency))

class_st <-
  stacks() %>%
  add_candidates(class_res_nn) %>%
  add_candidates(class_res_rf) %>%
  blend_predictions() %>%
  fit_members()
#> Warning: Predictions from 1 candidate were identical to those from existing candidates
#> and were removed from the data stack.

augment(class_st, tree_frogs_class)
#> # A tibble: 1,212 × 6
#>    treatment  reflex    age t_o_d     hatched .pred_class
#>    <chr>      <fct>   <dbl> <chr>     <chr>   <fct>      
#>  1 control    full   466965 morning   yes     full       
#>  2 gentamicin full   404310 afternoon no      full       
#>  3 gentamicin full   426220 night     no      full       
#>  4 control    mid    355360 night     no      mid        
#>  5 gentamicin mid    356535 night     no      mid        
#>  6 control    low    361180 night     yes     mid        
#>  7 gentamicin full   400070 afternoon no      full       
#>  8 control    full   401595 afternoon yes     full       
#>  9 control    mid    357810 night     yes     mid        
#> 10 control    mid    358410 night     no      mid        
#> # … with 1,202 more rows

augment(class_st, tree_frogs_class, type = "prob") %>%
  select(reflex, starts_with("."))
#> # A tibble: 1,212 × 4
#>    reflex .pred_full .pred_low .pred_mid
#>    <fct>       <dbl>     <dbl>     <dbl>
#>  1 full   0.999       0.000542  0.000258
#>  2 full   0.999       0.000659  0.000257
#>  3 full   0.998       0.00144   0.000326
#>  4 mid    0.000109    0.466     0.534   
#>  5 mid    0.0000477   0.323     0.677   
#>  6 low    0.0000175   0.254     0.746   
#>  7 full   0.999       0.000659  0.000257
#>  8 full   0.999       0.000513  0.000255
#>  9 mid    0.00000753  0.215     0.785   
#> 10 mid    0.000157    0.357     0.642   
#> # … with 1,202 more rows
```

<sup>Created on 2023-01-18 with [reprex v2.0.2](https://reprex.tidyverse.org)</sup>